+++ jamie/_meta.py
@@ -1,1 +1,1 @@
-__version__ = '4.3.0-pre1'
+__version__ = '4.3.0-pre2'